Reorganizing layout
