+++ tests/core/test_validate.py
@@ -4,7 +4,7 @@ from apistar import validators
 from apistar.exceptions import (
     ErrorMessage, ParseError, Position, ValidationError
 )
-from apistar.validate import validate
+from apistar.core import validate
 
 
 def test_valid_json():
